Modified `perspective-viewer-highcharts` to allow chart types to be imported individually.
