+++ packages/perspective-viewer-highcharts/package.json
@@ -2,12 +2,25 @@
     "name": "@jpmorganchase/perspective-viewer-highcharts",
     "version": "0.2.23",
     "description": "Perspective.js",
-    "main": "cjs/js/highcharts.js",
-    "browser": "cjs/js/highcharts.js",
-    "module": "cjs/js/highcharts.js",
+    "main": "./cjs/js/index.js",
+    "unpkg": "./build/highcharts.plugin.js",
+    "browser": {
+        "heatmap": "./cjs/js/heatmap.js",
+        "sunburst": "./cjs/js/sunburst.js",
+        "treemap": "./cjs/js/treemap.js",
+        "x_bar": "./cjs/js/x_bar.js",
+        "xy_line": "./cjs/js/xy_line.js",
+        "xy_scatter": "./cjs/js/xy_scatter.js",
+        "y_area": "./cjs/js/y_area.js",
+        "y_bar": "./cjs/js/y_bar.js",
+        "y_line": "./cjs/js/y_line.js",
+        "y_scatter": "./cjs/js/y_scatter.js"
+    },
+    "module": "./cjs/js/index.js",
     "files": [
         "build/**/*",
-        "cjs/**/*"
+        "cjs/**/*",
+        "babel.config.js"
     ],
     "scripts": {
         "bench": "npm-run-all bench:build bench:run",
