+++ packages/perspective-viewer-highcharts/src/js/highcharts/color_axis.js
