+++ packages/perspective-viewer-highcharts/src/js/highcharts/config.js
