+++ packages/perspective-viewer-highcharts/src/js/highcharts/series.js
