+++ packages/perspective-viewer-highcharts/src/js/highcharts/tooltip.js
