+++ packages/perspective-viewer-highcharts/src/js/sunburst.js
@@ -1,0 +1,11 @@
+/******************************************************************************
+ *
+ * Copyright (c) 2017, the Perspective Authors.
+ *
+ * This file is part of the Perspective library, distributed under the terms of
+ * the Apache License 2.0.  The full license can be found in the LICENSE file.
+ *
+ */
+
+import register from "./highcharts/highcharts.js";
+register("sunburst");
